Convert int24 frames (#593)
If AV_CODEC_ID_PCM_S24BE is used, it returns frames in 24 bits pixel formats in AV_SAMPLE_FMT_S32. This introduces scratching while playing.
Added explicit converting such frames to AV_SAMPLE_FMT_S32.

+++ tests/auto/integration/qavdemuxer/tst_qavdemuxer.cpp
@@ -161,7 +161,7 @@ void tst_QAVDemuxer::loadAudio()
         QVERIFY(af.stream().codec()->codec() != nullptr);
 
         auto format = af.format();
-        QCOMPARE(format.sampleFormat(), QAVAudioFormat::Int32);
+        QCOMPARE(format.sampleFormat(), QAVAudioFormat::Int16);
         auto data = af.data();
         QVERIFY(!data.isEmpty());
 
@@ -363,7 +363,7 @@ void tst_QAVDemuxer::qrcIO()
         QVERIFY(af.stream().codec()->codec() != nullptr);
 
         auto format = af.format();
-        QCOMPARE(format.sampleFormat(), QAVAudioFormat::Int32);
+        QCOMPARE(format.sampleFormat(), QAVAudioFormat::Int16);
         auto data = af.data();
         QVERIFY(!data.isEmpty());
